requirements.txt
speedtest-cli
pandas
matplotlib
requests
